+++ tests/lastfm-info-test.js
@@ -11,14 +11,6 @@ describe("a new info instance")
     gently = new Gently();
   });
 
-  it("accepts listeners in options (deprecated)", function() {
-      var handlers = { error: function() {}, success: function() {} };
-      gently.expect(handlers, "error");
-      gently.expect(handlers, "success");
-      var info = new LastFmInfo(lastfm, "", handlers);
-      info.emit("success");
-    });
-
   it("accepts listeners in handler options", function() {
     var handlers = { error: function() {}, success: function() {} };
     var options = { handlers: handlers };